chore(chat): split `models.reducers.ts` to types, selectors and reducers
