+++ apps/chat/src/store/models/models.types.ts
@@ -1,0 +1,23 @@
+import { ErrorMessage } from '@/src/types/error';
+import {
+  DialAIEntityModel,
+  InstalledModel,
+  ModelsMap,
+  PublishRequestDialAIEntityModel,
+} from '@/src/types/models';
+
+import { UploadStatus } from '@epam/ai-dial-shared';
+
+export interface ModelsState {
+  initialized: boolean;
+  status: UploadStatus;
+  error: ErrorMessage | undefined;
+  models: DialAIEntityModel[];
+  modelsMap: ModelsMap;
+  recentModelsIds: string[];
+  recentModelsStatus: UploadStatus;
+  isInstalledModelsInitialized: boolean;
+  installedModels: InstalledModel[];
+  publishRequestModels: PublishRequestDialAIEntityModel[];
+  publishedApplicationIds: string[];
+}
